rename analysis component
